backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
cryptography==41.0.7
supabase==2.3.1
python-multipart==0.0.6
httpx==0.25.2